requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
python-decouple>=3.8
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-magic>=0.4.27
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
django-cors-headers>=4.3.0
drf-yasg2>=1.19.4

djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10

# uWSGI (for Docker + supervisord deployment)
uWSGI==2.0.26
